// themes/theme-b2b/src/colors.d.ts
export declare const colors: {
  gray00: string;
  gray10: string;
  gray20: string;
  gray30: string;
  gray40: string;
  gray50: string;
  gray60: string;
  gray70: string;
  gray80: string;
  gray90: string;
  orange10: string;
  orange20: string;
  orange40: string;
  orange60: string;
  orange70: string;
  orange80: string;
  blue10: string;
  blue20: string;
  blue40: string;
  blue60: string;
  blue70: string;
  blue80: string;
  green10: string;
  green20: string;
  green40: string;
  green60: string;
  green70: string;
  green80: string;
  red10: string;
  red20: string;
  red40: string;
  red60: string;
  red70: string;
  red80: string;
  yellow10: string;
  yellow30: string;
  yellow40: string;
  yellow60: string;
  yellow70: string;
  yellow80: string;
};
//# sourceMappingURL=colors.d.ts.map